Make SetRayTracingShaderGroupHandlesCommandHeader`s size constant
One of the members of SetRayTracingShaderGroupHandlesCommandHeader was
of size_t type. This would cause trouble when replaying captures
captured in 32bit with a 64bit gfxrecon-replay or vice versa

This commit converts that member variable to uint64_t in order to have a
fixed size

+++ framework/format/format.h
@@ -456,7 +456,7 @@ struct SetRayTracingShaderGroupHandlesCommandHeader
     format::ThreadId thread_id;
     format::HandleId device_id;
     format::HandleId pipeline_id;
-    size_t           data_size;
+    uint64_t         data_size;
 };
 
 struct CreateHeapAllocationCommand
